+++ components/Layouts/Auth/index.jsx
@@ -1,16 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-const Auth = ({ children, hideSideBarOnMobile = false }) => (
+const Auth = ({ title = '', children, hideSideBarOnMobile = false }) => (
   <div className="w-full flex flex-col md:flex-row max-w-screen-xl min-h-screen mx-auto">
     <div
       className={`bg-tertiary px-5 pt-12 pb-8 w-360 md:flex-shrink-0 md:pt-24 md:relative md:overflow-y-hidden ${
         hideSideBarOnMobile ? 'hidden md:block' : 'block'
       }`}
     >
-      <h3 className="text-center md:text-left md:max-w-xs">
-        Join Scizzor & Increase Your Instore Appointments
-      </h3>
+      <h3 className="text-center md:text-left md:max-w-xs">{title}</h3>
       <img
         className="hidden md:block w-80 h-80 -ml-10 -mb-16 object-cover rounded-full absolute bottom-0 left-0"
         src="/images/auth.jpg"
@@ -26,6 +24,7 @@ const Auth = ({ children, hideSideBarOnMobile = false }) => (
 Auth.propTypes = {
   children: PropTypes.node.isRequired,
   hideSideBarOnMobile: PropTypes.bool,
+  title: PropTypes.oneOfType([PropTypes.string, PropTypes.node]).isRequired,
 };
 
 export default Auth;
