+++ components/Layouts/Auth/index.test.js
@@ -4,5 +4,5 @@ import { shallow } from 'enzyme';
 import Auth from './index';
 
 it('renders without crashing', () => {
-  shallow(<Auth>test</Auth>);
+  shallow(<Auth title="Test">test</Auth>);
 });
