+++ components/index.js
@@ -4,4 +4,4 @@ export { default as Auth } from './Layouts/Auth';
 
 // forms
 export { default as BrandSignUpForm } from './Forms/BrandSignUpForm';
-export { default as BrandSignInForm } from './Forms/BrandSignInForm';
+export { default as SignInForm } from './Forms/SignInForm';
